comments

+++ src/main/java/net/offbeatpioneer/retroengine/core/util/ColorTools.java
@@ -15,15 +15,15 @@ public class ColorTools {
      *
      * @param color1    first colour
      * @param color2    second colour
-     * @param tolerance tolerance
+     * @param tolerance tolerance, between 0 - 255 (RGB range)
      * @return true, if the colours matches within the tolerance, otherwise false
      */
     public static boolean closeMatch(int color1, int color2, int tolerance) {
-        if ((int) Math.abs(Color.red(color1) - Color.red(color2)) > tolerance)
+        if (Math.abs(Color.red(color1) - Color.red(color2)) > tolerance)
             return false;
-        if ((int) Math.abs(Color.green(color1) - Color.green(color2)) > tolerance)
+        if (Math.abs(Color.green(color1) - Color.green(color2)) > tolerance)
             return false;
-        if ((int) Math.abs(Color.blue(color1) - Color.blue(color2)) > tolerance)
+        if (Math.abs(Color.blue(color1) - Color.blue(color2)) > tolerance)
             return false;
         return true;
     }
